불필요한 파일들 삭제
