+++ src/nd01.py
@@ -20,7 +20,7 @@ from PyQt5.QtMultimediaWidgets import QGraphicsVideoItem
 from video_thread import VideoThread
 from curved import CurvedThread
 from mainwindow import Ui_MainWindow
-from video_test import Js06VideoWidget2
+from video_thread import Js06VideoWidget2
 
 print(pd.__version__)
